+++ verifier-application/src/test/java/ch/admin/bj/swiyu/verifier/DefaultExceptionHandlerTest.java
@@ -34,8 +34,11 @@ class DefaultExceptionHandlerTest {
 
         // Then
         assertEquals(HttpStatus.BAD_REQUEST, response.getStatusCode());
+
+        var apiErrorDto = (ApiErrorDto) response.getBody();
+
         assertNotNull(response.getBody());
-        assertEquals(errorMessage, response.getBody());
+        assertEquals(errorMessage, apiErrorDto.getErrorDescription());
     }
 
     @Test
@@ -45,9 +48,9 @@ class DefaultExceptionHandlerTest {
         ResponseEntity<Object> response = handler.handleIllegalArgumentException(ex);
 
         // Then
+        var apiErrorDto = (ApiErrorDto) response.getBody();
         assertEquals(HttpStatus.BAD_REQUEST, response.getStatusCode());
-        assertNotNull(response.getBody());
-        assertEquals("Bad request", response.getBody());
+        assertEquals("Bad request", apiErrorDto.getErrorDescription());
     }
 
     @Test
@@ -58,9 +61,10 @@ class DefaultExceptionHandlerTest {
         ResponseEntity<Object> response = handler.handleNoSuchElementException(ex);
 
         // Then
+        var apiErrorDto = (ApiErrorDto) response.getBody();
         assertEquals(HttpStatus.NOT_FOUND, response.getStatusCode());
         assertNotNull(response.getBody());
-        assertEquals(errorMessage, response.getBody());
+        assertEquals(errorMessage, apiErrorDto.getErrorDescription());
     }
 
     @Test
@@ -71,12 +75,11 @@ class DefaultExceptionHandlerTest {
         when(request.getRequestURL()).thenReturn(new StringBuffer("http://localhost/test"));
 
         ResponseEntity<Object> response = handler.handleException(ex, request);
-
-        // Then
+        
         assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, response.getStatusCode());
         ApiErrorDto body = (ApiErrorDto) response.getBody();
         assertNotNull(response.getBody());
-        assertEquals("Internal Server Error. Please check again later", body.detail());
+        assertEquals("Internal Server Error. Please check again later", body.getErrorDescription());
     }
 
     @Test